requirements.txt
django-bower>=5.2.0
django-crispy-forms>=1.6.1
django-custom-user>=0.7
django-debug-toolbar>=1.6
django-extensions>=1.7.6
django-filter>=1.0.4
Django>=1.11.0
djangorestframework>=3.6.3
Pillow>=4.1.0
psycopg2>=2.7.1

-e git+https://github.com/basilfx/django-oauth2-provider.git#egg=django-oauth2-provider
